+++ molo/core/content_import/helpers/get_image.py
